requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Deprecated==1.2.14
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
idna==3.6
jwcrypto==1.5.0
Markdown==3.5.1
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
wrapt==1.16.0
